requirements_for_tests.txt
PyHamcrest
nose
mock
pep8==1.6.2
coverage
freezegun
